backendapp/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
flask-jwt-extended==4.4.4
supabase==1.0.0
psycopg2-binary==2.9.7
werkzeug==2.2.3
flask-bcrypt==1.0.1
python-dateutil==2.8.2
gunicorn==20.1.0
